Refactor mock to simplify Curator constructors
